renaming datasources to .config.json
